pyproject.toml
[project]
name = "news-db"
version = "0.1.0"
description = "Automated news fetching, processing, and storage system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.10.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "fetchers", "processors", "storage", "db"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
]
